fix(sana-wm/realtime): stop Module.to(dtype) from destroying complex RoPE buffers
ROOT CAUSE of the realtime<->batch stage-1 divergence (~54% relRMS on chunk 0 with
bitwise-identical inputs): SanaWMRealtimeStage.forward did
    transformer = self.transformer.to(device=device, dtype=weight_dtype)
and Module.to(dtype=...) also casts COMPLEX buffers (WanRotaryPosEmbed._freqs) to the
real weight dtype, silently discarding the imaginary part (torch UserWarning 'Casting
complex values to real discards the imaginary part') -> every RoPE rotation in the
realtime path ran with realified freqs. The batch path consumes the module as loaded
(use_declared_component moves device-only), so only realtime was corrupted.

Evidence (parity probes, 321-frame config):
- all 871 DiT parameters fingerprint-identical across paths
- chunk-0 step-0 noise_pred relRMS 40% pre-fix with bitwise-identical
  hidden_states/timestep/embeds/camera/plucker inputs
- post-fix: noise_pred 1.36%, chunk-0 stage-1 54.57% -> 1.17%, warning gone;
  residual matches the measured cond-frame encode delta (0.29% -> ~1.3%) exactly

Fix: device-only move; the loader already produces dit_precision weights.
Also adds parity probes (env-gated): per-step chunk-0 noise_pred dumps + DiT
weights fingerprint on both paths.

+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/model_specific_stages/sana_wm/sana_wm_realtime.py
@@ -939,7 +939,16 @@ class SanaWMRealtimeStage(PipelineStage):
         device = get_local_torch_device()
         weight_dtype = PRECISION_TO_TYPE[server_args.pipeline_config.dit_precision]
         vae_dtype = PRECISION_TO_TYPE[server_args.pipeline_config.vae_precision]
-        transformer = self.transformer.to(device=device, dtype=weight_dtype).eval()
+        # Device move WITHOUT a blanket dtype cast: Module.to(dtype=...) also casts
+        # the DiT's complex RoPE buffers (WanRotaryPosEmbed._freqs, complex) to the
+        # real weight dtype, silently DISCARDING the imaginary part (torch warns
+        # "Casting complex values to real discards the imaginary part") and
+        # corrupting every RoPE rotation. This was the realtime<->batch stage-1
+        # divergence root cause: bitwise-identical inputs + identical weights but
+        # ~40% relRMS on the very first forward. The loader already produces
+        # dit_precision weights, exactly as the batch streaming path consumes them
+        # (use_declared_component moves the module without re-casting).
+        transformer = self.transformer.to(device=device).eval()
         self.vae = self.vae.to(device=device, dtype=vae_dtype).eval()
         batch.enable_sequence_shard = int(server_args.sp_degree or 1) > 1
